Modify the way test names are created
string statements are pushed onto a name stack as they are encountered. When an
assertion site is encountered, the top name is popped off the stack and applied
as the name of the assertion site. Any existing names become the suite hierarchy
that the assertion site belongs to.

Add check for await expressions

Check for await expressions to determine whether
the source should be parsed by `Function` or
`AsyncFunction`

Rename compileTestSite to compileTestRig